Added Safari Support (disables IndexedDB, we don't have a cache sadly)

+++ src/common/refreshData.ts
@@ -1,14 +1,11 @@
 import { db, Icon } from '../data/IconsCache';
 import * as config from '../../res/config.json';
 import { DataStoreEvents, emitEvent, registerEvent } from "./eventmanager";
-
-registerEvent(DataStoreEvents.RefreshData, async (hmsys) =>
-{
+export let lastFilesCollected: any | undefined = undefined;
+registerEvent(DataStoreEvents.RefreshData, async (hmsys) => {
     const { data } = await hmsys.api.triggerResponse(config.targetId, { type: "getFiles" }) as any;
-
     const oldData = await db.icons.toArray();
-    const toUpdate: string[] = data.files.filter((x: Icon) =>
-    {
+    const toUpdate: string[] = data.files.filter((x: Icon) => {
         const thing = oldData.find(y => y.id == x.id)
         if (thing == undefined) return false;
         if (
@@ -26,21 +23,30 @@ registerEvent(DataStoreEvents.RefreshData, async (hmsys) =>
     const toAddBlobs = await Promise.all(toAddFetch.map(x => x.blob()))
     const toUpdateFetch: Response[] = await Promise.all(toUpdate.map((x: string) => hmsys.rest.get(config.targetId, `file/${x}`)))
     const toUpdateBlobs = await Promise.all(toUpdateFetch.map(x => x.blob()))
+    lastFilesCollected = toAdd.map((entry, index) => ({
+        ...data.files.find((x: Icon) => x.id == entry),
+        data: toAddBlobs[ index ]
+    }))
+    if (!/apple/i.test(navigator.vendor)) {
+        try {
+            await db.transaction('rw', db.icons, async () => {
+                await db.icons.bulkDelete(toRemove)
+                await db.icons.bulkPut(toUpdate.map((entry, index) =>
+                ({
+                    ...data.files.find((x: Icon) => x.id == entry),
+                    data: toUpdateBlobs[ index ]
+                })))
+                await db.icons.bulkAdd(toAdd.map((entry, index) =>
+                ({
+                    ...data.files.find((x: Icon) => x.id == entry),
+                    data: toAddBlobs[ index ]
+                })))
+            })
+        } catch (error) {
+            console.log(error)
+        }
 
-    await db.transaction('rw', db.icons, async () =>
-    {
-        await db.icons.bulkDelete(toRemove)
-        await db.icons.bulkPut(toUpdate.map((entry, index) =>
-        ({
-            ...data.files.find((x: Icon) => x.id == entry),
-            data: toUpdateBlobs[ index ]
-        })))
-        await db.icons.bulkAdd(toAdd.map((entry, index) =>
-        ({
-            ...data.files.find((x: Icon) => x.id == entry),
-            data: toAddBlobs[ index ]
-        })))
-    })
+    }
     if (toAdd.length !== 0 || toRemove.length !== 0 || toUpdate.length !== 0)
         emitEvent(DataStoreEvents.RefreshDataComplete, { new: toAdd, removed: toRemove, updated: toUpdate })
 })
